dataAddNewTransactionSignature
